password-hash: extract `phc` submodule
Extracts all functionality related to the Password Hashing Competition
(PHC) string format into a `phc` submodule.

The longer-term goal would be to extract this into its own crate,
however that will need some trait redesign, which can be better
accomplished by first adding a `phc` feature and feature-gating the
relevant functionality.

This is just an initial step to begin isolating the relevant code.
