+++ password-hash/src/phc/ident.rs
